FEATBL-1789 Feat: add methods RETR, UIDL, TOP, DEL

+++ parent/mailbackend/net.bluemind.backend.mail.replica.service/src/net/bluemind/backend/mail/replica/service/internal/hooks/CleanTemporaryParts.java
@@ -19,10 +19,11 @@ public class CleanTemporaryParts implements ISessionDeletionListener {
 		File sidFolder = Bodies.getFolder(sid);
 
 		File[] parts = sidFolder.listFiles();
-		for (File part : parts) {
-			part.delete();
+		if (parts != null) {
+			for (File part : parts) {
+				part.delete();
+			}
 		}
-
 		sidFolder.delete();
 
 		logger.debug("Folder {} deleted.", sidFolder.getAbsolutePath());
